remove layout.selectors from job parameters (#109010)
