+++ x-pack/plugins/reporting/public/index.ts
@@ -6,20 +6,18 @@
  */
 
 import { PluginInitializerContext } from 'src/core/public';
-import { getDefaultLayoutSelectors } from '../common';
 import { ReportingAPIClient } from './lib/reporting_api_client';
 import { ReportingPublicPlugin } from './plugin';
 import { getSharedComponents } from './shared';
 
 export interface ReportingSetup {
-  getDefaultLayoutSelectors: typeof getDefaultLayoutSelectors;
   usesUiCapabilities: () => boolean;
   components: ReturnType<typeof getSharedComponents>;
 }
 
 export type ReportingStart = ReportingSetup;
 
-export { constants, getDefaultLayoutSelectors } from '../common';
+export { constants } from '../common';
 export { ReportingAPIClient, ReportingPublicPlugin as Plugin };
 
 export function plugin(initializerContext: PluginInitializerContext) {
