+++ x-pack/plugins/reporting/public/mocks.ts
@@ -8,7 +8,6 @@
 import { coreMock } from 'src/core/public/mocks';
 import { ReportingAPIClient } from './lib/reporting_api_client';
 import { ReportingSetup } from '.';
-import { getDefaultLayoutSelectors } from '../common';
 import { getSharedComponents } from './shared';
 
 type Setup = jest.Mocked<ReportingSetup>;
@@ -17,7 +16,6 @@ const createSetupContract = (): Setup => {
   const coreSetup = coreMock.createSetup();
   const apiClient = new ReportingAPIClient(coreSetup.http, coreSetup.uiSettings, '7.15.0');
   return {
-    getDefaultLayoutSelectors: jest.fn().mockImplementation(getDefaultLayoutSelectors),
     usesUiCapabilities: jest.fn().mockImplementation(() => true),
     components: getSharedComponents(coreSetup, apiClient),
   };
