+++ x-pack/plugins/reporting/server/export_types/common/generate_png.ts
@@ -32,7 +32,7 @@ export async function generatePngObservableFactory(reporting: ReportingCore) {
     if (!layoutParams || !layoutParams.dimensions) {
       throw new Error(`LayoutParams.Dimensions is undefined.`);
     }
-    const layout = new PreserveLayout(layoutParams.dimensions, layoutParams.selectors);
+    const layout = new PreserveLayout(layoutParams.dimensions);
     if (apmLayout) apmLayout.end();
 
     const apmScreenshots = apmTrans?.startSpan('screenshots_pipeline', 'setup');
