+++ x-pack/plugins/reporting/server/lib/layouts/index.ts
@@ -6,21 +6,22 @@
  */
 
 import { LevelLogger } from '../';
-import { LayoutSelectorDictionary, Size } from '../../../common/types';
+import { Size } from '../../../common/types';
 import { HeadlessChromiumDriver } from '../../browsers';
 import type { Layout } from './layout';
 
-export {
-  LayoutParams,
-  LayoutSelectorDictionary,
-  PageSizeParams,
-  PdfImageSize,
-  Size,
-} from '../../../common/types';
+export interface LayoutSelectorDictionary {
+  screenshot: string;
+  renderComplete: string;
+  itemsCountAttribute: string;
+  timefilterDurationAttribute: string;
+}
+
+export { LayoutParams, PageSizeParams, PdfImageSize, Size } from '../../../common/types';
+export { CanvasLayout } from './canvas_layout';
 export { createLayout } from './create_layout';
 export type { Layout } from './layout';
 export { PreserveLayout } from './preserve_layout';
-export { CanvasLayout } from './canvas_layout';
 export { PrintLayout } from './print_layout';
 
 export const LayoutTypes = {
