+++ x-pack/plugins/reporting/server/lib/layouts/preserve_layout.ts
@@ -7,33 +7,28 @@
 
 import path from 'path';
 import { CustomPageSize } from 'pdfmake/interfaces';
-import { getDefaultLayoutSelectors } from '../../../common';
 import { LAYOUT_TYPES } from '../../../common/constants';
-import { LayoutSelectorDictionary, PageSizeParams, Size } from '../../../common/types';
-import type { LayoutInstance } from './';
+import { PageSizeParams, Size } from '../../../common/types';
+import { getDefaultLayoutSelectors, LayoutInstance } from './';
 import { Layout } from './layout';
 
 // We use a zoom of two to bump up the resolution of the screenshot a bit.
 const ZOOM: number = 2;
 
 export class PreserveLayout extends Layout implements LayoutInstance {
-  public readonly selectors: LayoutSelectorDictionary = getDefaultLayoutSelectors();
+  public readonly selectors = getDefaultLayoutSelectors();
   public readonly groupCount = 1;
   public readonly height: number;
   public readonly width: number;
   private readonly scaledHeight: number;
   private readonly scaledWidth: number;
 
-  constructor(size: Size, layoutSelectors?: LayoutSelectorDictionary) {
+  constructor(size: Size) {
     super(LAYOUT_TYPES.PRESERVE_LAYOUT);
     this.height = size.height;
     this.width = size.width;
     this.scaledHeight = size.height * ZOOM;
     this.scaledWidth = size.width * ZOOM;
-
-    if (layoutSelectors) {
-      this.selectors = layoutSelectors;
-    }
   }
 
   public getCssOverridesPath() {
